+++ AOS/app/src/main/java/org/softeer/robocar/data/dto/route/response/RouteSoloResponse.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class RouteSoloResponse(
     @SerialName("carImage") val carImage: String,
-    @SerialName("estimatedArrivalTime") val estimatedArrivalTime: Long,
+    @SerialName("estimatedArrivalTime") val estimatedArrivalTime: Int,
     @SerialName("carNumber") val carNumber: String,
     @SerialName("carName") val carName: String,
     @SerialName("nodes") val nodes: List<Coordinate>
