+++ AOS/app/src/main/java/org/softeer/robocar/data/model/RouteSolo.kt
@@ -7,7 +7,7 @@ import org.softeer.robocar.data.dto.route.response.Coordinate
 @Serializable
 data class RouteSolo(
     val carImageUrl: String,
-    val estimatedArrivalTime: Long,
+    val estimatedArrivalTime: Int,
     val vehicleRegistrationNumber: String,
     val vehicleModel: String,
     val nodes: List<Coordinate>
